% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.strict_reuse_warning}
\alias{.strict_reuse_warning}
\title{Return a warning string when N > ratio*U (strict >) suggesting a lighter mode, else NULL.}
\usage{
.strict_reuse_warning(U, N, ratio)
}
\description{
Return a warning string when N > ratio*U (strict >) suggesting a lighter mode, else NULL.
}
\keyword{internal}
